Adds expense creation related specs
As of this commit expense creation in the app happens based on two
scenarios, namely -
1. When a user has no expenses at all
2. When a user has atleast one expense

If you observe closely, you may find that for scenario 2 there are no
specs for the valid and invalid information checks when going about
creating an expense. This is basically because scenario 2 goes to the
same page to create a new expense as is in the case of scenario 1. Thus
in scenario 2 it is sufficient if we test that a user goes to same
page(as the one used in case of scenario 1) where one can go about
creating a new expense.