Adds expenses updation related specs
There are a couple of things to note here.

Before we could delve into talking about the code changes, I’d like to
first add that only when an expense is created that we go about
actually seeing the ‘Edit’ link wrt user home page. Let’s call this as
point 1.

Now coming to the code related changes -
1. We’ve made use of let with a bang(!) because we can’t afford the
expense to be lazy loaded. An expense has to be created before hand so
that we could make use of the ‘Edit’ link as mentioned in point 1.

2. Since we’re actually creating an expense from the backend we need to
actually refresh the user home page. Hence we’ve added the code related
to ‘visit root_path’. Only with this line of code we’ll see the ‘Edit’
link.

3. When updating an expense we’re basically considering two scenarios -
a. With invalid expense information
b. With valid expense information