Adds user password update related specs
For one to be able to send a request to update a user’s password, one
needs to be logged in to the system. Normally, as part of a user login
the session variable gets set and this currently happens within the
create action of the sessions controller. In RSpec we can’t call other
controller actions as of the details mentioned in this link -
http://stackoverflow.com/questions/7290035/post-to-a-different-controlle
r-in-an-rspec-test, hence we can’t hit sessions#create from within
users controller specs to make a user login. Hence, we’re manually
logging in the user by setting the session and this is what we’ll be
doing when testing all non integration tests.