Uses poltergeist and database_cleaner to test JS dependent features
Spawning a browser window each time is slow and hence here we use
Poltergeist(a ruby wrapper for PhantomJS - a solution which facilitates
headless website testing) here. Also, we need to register the
poltergeist driver with Capybara in order to make this work.

Lastly, javascript related tests cannot be used with transactional
fixtures. With database cleaner we can switch to using a truncation
strategy when testing js related features.