Ensures the date is converted to appropriate format
As part of the total expenses form, the form params are passed as a
string. Inorder to do comparison between two dates to validate things
like the from_date should be less than the to_date in order to
calculate expenses within a period, the dates need to be in the
appropriate format before the comparison that validates this. Hence
this conversion is done .