Handles a scenario where a user tries to delete an already deleted expense
Scenario: Assume a user is logged in to the expense app from two
different browsers and the user deletes one of their expenses from one
of the browsers. In case the user forgets about this action and is
surprised to see the same record from the other browser and goes about
trying to delete the same record an appropriate error message is
displayed as part of this user action.