Ensures that the DB index is created on the right set of columns
User expenses are displayed on app homepage. Adding the multiple key DB
index on “date” and “user_id” helps render the app homepage faster as
the expenses on this page are ordered based on the “date”(the date when
an expenses was incurred) of an actual expense and not based on when
the expense record was “created_at” wrt the app, hence appropriate
corrections are made here. For more
details on how this could be helpful, feel free to refer to the link -
http://stackoverflow.com/questions/4445507/rails-created-at-when-user-fo
r-ordering-should-you-add-an-index-to-the-table