Adds flash message related spec if a logged in user tries to access login page
A logged in user shouldn’t have access to the login page. If a user
tries to do so, they are redirected back to the app home page along
with the display of an appropriate error message. This commit ensures
that the correct error message gets displayed to a user wrt the above
mentioned scenario.