Moves integration tests from requests to features
As mentioned in the blog -
http://www.andylindeman.com/rspec-rails-and-capybara-2.0-what-you-need-t
o-know/, as of Capybara 2.0 and above, capybara by itself would be
available as part of feature specs by default. As part of this commit
we remove the explicit inclusion of the Capybara DSL as this is not a
recommended approach as mentioned in the link -
https://github.com/rspec/rspec-rails/blob/master/Capybara.md.

Also as part of this change we’d be able to differentiate if a given
test is an integration test or not if we check if the Capybara method
“visit” is defined or not. This is usually defined when one uses them
as part of feature specs.

Lastly, this commit is a WIP in terms of the overall effort of moving
all request specs over to feature specs. Some of the request specs wrt
authentication_pages_spec needs to be modified for them to be fully
working as part of feature specs.