Adds changes around deletion specs along with other misc changes
This commit has:
1. Below changes to deletion specs -
a. Ensures that deletion specs use xhr to submit an AJAX request for
expense deletion
b. Ensures that an appropriate error message is raised if a user tries
to delete the same expense twice. This replaces the original message.
c. Use flash messages throughout for AJAX and HTML submit requests
d. For an AJAX request submit if an expense isn’t there, with the
current code one cannot redirect to a different page, hence this case
is accordingly handled by just showing the error message.
2. Misc changes that include -
- A more generic error message is raised